+++ binless/R/binless.R
@@ -29,20 +29,19 @@
 #'
 #' @useDynLib binless, .registration = TRUE
 #' @import rstan
-#' @import rstantools
 #' @import methods
-#' @import splines
 #' @import data.table
 #' @import doParallel
 #' @import foreach
 #' @import matrixStats
 #' @import ggplot2
-#' @import scales
 #' @import MASS
 #' @import Matrix
 #' @import quadprog
+#' @importFrom scales muted
 #' @importFrom Hmisc cut2
 #' @importFrom dplyr ntile
+#' @importFrom utils packageVersion
 #'
 #' @docType package
 #' @name binless