+++ binless/R/detect_binless.R
@@ -3,7 +3,7 @@ NULL
 
 #' Prepare grouped signal matrix and settings
 #' @keywords internal
-prepare_signal_estimation = function(cs, csg, resolution, tol.val) {
+prepare_signal_estimation = function(cs, csg, resolution, tol.val, nperf) {
   mat = binless:::get_signal_matrix(cs, resolution, groups=csg@names)
   #
   #add trail information
@@ -14,9 +14,7 @@ prepare_signal_estimation = function(cs, csg, resolution, tol.val) {
                 nbins = csg@par$nbins,
                 dispersion = csg@par$alpha,
                 tol.val = tol.val,
-                nperf = 50,
-                min.patchsize = 4,
-                min.l10FC = 0.5)
+                nperf = nperf)
   cts=csg@cts[,.(name,bin1,bin2,count,lmu.nosig,weight,log_decay)]
   csi=new("CSbsig", mat=mat, cts=cts, settings=settings)
   return(csi)
@@ -24,8 +22,8 @@ prepare_signal_estimation = function(cs, csg, resolution, tol.val) {
 
 #' Build grouped difference matrix using normalization data if available
 #' @keywords internal
-prepare_difference_estimation = function(cs, csg, resolution, ref, tol.val) {
-  csi = binless:::prepare_signal_estimation(cs, csg, resolution, tol.val)
+prepare_difference_estimation = function(cs, csg, resolution, ref, tol.val, nperf) {
+  csi = binless:::prepare_signal_estimation(cs, csg, resolution, tol.val, nperf)
   names=csg@names
   mat = foreach(n=names[groupname!=ref,unique(groupname)],.combine=rbind) %do%
     merge(csi@mat[name==n],csi@mat[name==ref,.(bin1,bin2,phi1=phi)],all=T,by=c("bin1","bin2"))
@@ -40,19 +38,22 @@ prepare_difference_estimation = function(cs, csg, resolution, ref, tol.val) {
 
 #' Perform binless interaction detection using fused lasso
 #'
-#' @param cs 
-#' @param ref 
-#' @param resolution 
-#' @param group 
-#' @param ncores 
-#' @param niter number of IRLS iterations, and BIC iterations within
+#' @param cs a normalized CSnorm object
+#' @param resolution The base resolution (default: the one used during normalization)
+#' @param group Default "all", but also various groupings, see \code{\link{group_datasets}}
+#' @param ncores How many cores to parallelize on (default: the one used during normalization)
+#' @param verbose Whether to be verbose (default) or not
+#' @param tol.val Tolerance on computed quantities
+#' @param fix.lambda1,fix.lambda1.at,fix.lambda2,fix.lambda2.at Default FALSE. if TRUE, sets lambda1
+#'  (resp. lambda2) to the provided value instead of optimizing it
+#' @param nperf number of IRLS iterations (default 50)
 #'   
 #' @return 
 #' @export
 #' 
 #' @examples
 detect_binless_interactions = function(cs, resolution=cs@settings$base.res, group="all", ncores=1, tol.val=cs@settings$tol, verbose=T,
-                                       fix.lambda1=F, fix.lambda1.at=NA, fix.lambda2=F, fix.lambda2.at=NA){
+                                       fix.lambda1=F, fix.lambda1.at=NA, fix.lambda2=F, fix.lambda2.at=NA, nperf=50){
   if (verbose==T) cat("Binless interaction detection with resolution=",resolution," and group=",group,"\n")
   ### get CSgroup object
   idx1=get_cs_group_idx(cs, resolution, group, raise=T)
@@ -63,7 +64,7 @@ detect_binless_interactions = function(cs, resolution=cs@settings$base.res, grou
   #
   ### prepare signal estimation
   if (verbose==T) cat("  Prepare for signal estimation\n")
-  csi = binless:::prepare_signal_estimation(cs, csg, resolution, tol.val)
+  csi = binless:::prepare_signal_estimation(cs, csg, resolution, tol.val, nperf)
   #
   #perform fused lasso on signal
   if (verbose==T) cat("  Fused lasso\n")
@@ -101,6 +102,7 @@ detect_binless_interactions = function(cs, resolution=cs@settings$base.res, grou
 #' Binless detection of significant differences with a reference
 #' 
 #' @inheritParams detect_binless_interactions
+#' @param ref The reference to compute the difference on, as obtained from cs@experiments
 #'   
 #' @return the binned matrix with additional information relating to these
 #'   significant interactions
@@ -108,7 +110,7 @@ detect_binless_interactions = function(cs, resolution=cs@settings$base.res, grou
 #' 
 #' @examples
 detect_binless_differences = function(cs, ref, resolution=cs@settings$base.res, group="all", ncores=1, tol.val=cs@settings$tol, verbose=T,
-                                      fix.lambda1=F, fix.lambda1.at=NA, fix.lambda2=F, fix.lambda2.at=NA){
+                                      fix.lambda1=F, fix.lambda1.at=NA, fix.lambda2=F, fix.lambda2.at=NA, nperf=50){
   if (verbose==T) cat("Binless difference detection with resolution=",resolution,
                       " group=", group," and ref=",as.character(ref),"\n")
   ### get CSgroup object
@@ -119,7 +121,7 @@ detect_binless_differences = function(cs, ref, resolution=cs@settings$base.res, 
   if (is.character(ref)) ref=csg@names[as.character(groupname)==ref,unique(groupname)]
   if (length(ref) == 0) stop(paste0("Invalid ref! For group by ",group,", acceptable refs are: ",paste(levels(ref),collapse=" / ")))
   if (verbose==T) cat("  Prepare for difference estimation\n")
-  csi = binless:::prepare_difference_estimation(cs, csg, resolution, ref, tol.val)
+  csi = binless:::prepare_difference_estimation(cs, csg, resolution, ref, tol.val, nperf)
   #
   #perform fused lasso on signal
   if (verbose==T) cat("  Fused lasso\n")