+++ binless/src/IRLSEstimator.hpp
@@ -18,7 +18,7 @@ public:
     //requesting precision to be below a given convergence criterion
     //final beta value will be clamped if clamp>0
     IRLSEstimator(double converge, GaussianEstimator& gauss, WeightsUpdater& wt)
-    : Timer("IRLSEstimator"), converge_(converge), counter_(0), gauss_(gauss), wt_(wt) {}
+    : Timer("IRLSEstimator"), converge_(converge), counter_(0), has_converged_(false), gauss_(gauss), wt_(wt) {}
     
     //run the optimization on the given data. The objective is
     // sum_i w_i(y_i-beta_i)^2 + lambda2 * sum_ij |beta_i-beta_j|
@@ -31,7 +31,7 @@ public:
         double precision = converge_+1;
         std::vector<double> beta_old = beta;
         /*Rcpp::Rcout << " Perf iteration: start with lam2= " << lambda2 << " alpha= "
-                    << gauss_.get_alpha() << " phi[0]= " << beta[0] << "\n";*/
+                      << gauss_.get_alpha() << " phi[0]= " << beta[0] << "\n";*/
         do {
           //update weights
           Timer::start_timer("weight");
@@ -44,14 +44,15 @@ public:
           beta = gauss_.get();
           //update counters and compute precision
           precision = get_precision(beta,beta_old);
+          has_converged_ = precision < converge_;
           ++counter_;
           beta_old = beta;
           /*Rcpp::Rcout << " Iteration " << counter_ << " / " << nouter << " with lam2= " << lambda2 << " alpha= "
             << gauss_.get_alpha() << " reached maxval= " << precision
             << " after " << gauss_.get_ninner() << " steps " << " phi[0]= " << beta[0] << "\n";*/
-        } while (counter_ < nouter && precision > converge_ );
+        } while (counter_ < nouter && (!has_converged_) );
         /*Rcpp::Rcout << " Perf iteration: end with lam2= " << lambda2 << " alpha= "
-        << gauss_.get_alpha() << " phi[0]= " << beta[0] << "\n";*/
+                      << gauss_.get_alpha() << " phi[0]= " << beta[0] << "\n";*/
         Timer::start_timer("weight");
         wt_.set_beta(beta); //store last beta without updating weights (but might update phi_ref if applicable)
         Timer::print_times();
@@ -61,6 +62,11 @@ public:
     unsigned get_nouter() const {
         return counter_;
     }
+  
+  //return the number of outer iterations
+  bool has_converged() const {
+    return has_converged_;
+  }
     
 private:
     double get_precision(const std::vector<double>& beta, const std::vector<double>& beta_old) const {
@@ -74,6 +80,7 @@ private:
         
     const double converge_;
     unsigned counter_;
+    bool has_converged_;
     
     GaussianEstimator& gauss_;
     WeightsUpdater& wt_;